offst-test: Initial work splitting node_report_service to a separate file.
