+++ components/test/src/lib.rs
@@ -49,6 +49,9 @@ mod app_wrapper;
 mod compact_node_wrapper;
 
 #[cfg(test)]
+mod node_report_service;
+
+#[cfg(test)]
 mod tests;
 
 #[cfg(test)]
